module_3(fix): functions refactoring
