+++ src/utils/const/const.ts
@@ -1,3 +1,4 @@
 export const RATES_KEY = "5cdeb4588a548a8b25607bae";
 export const NEWS_KEY = "df30d7c635e24a55800b4adfb6b0aafb";
 export const UPDATE_TIME = 1000 * 60 * 15;
+export const BASE_URL = "http://localhost:8080";
