+++ src/utils/functions/news/checkNewsItem.ts
