+++ src/utils/functions/removeSpaces.ts
@@ -1,0 +1,2 @@
+export const removeSpaces = (string: string) =>
+  string.replace(/\s+/g, "").trim();
